Switch to off-heap PageList
The PageCacheTests pass, but the PageCacheSlowIT do not. Yet.

+++ community/io/src/test/java/org/neo4j/adversaries/pagecache/AdversarialPageCache.java
@@ -109,7 +109,7 @@ public class AdversarialPageCache implements PageCache
     }
 
     @Override
-    public int maxCachedPages()
+    public long maxCachedPages()
     {
         return delegate.maxCachedPages();
     }
